+++ tests/EndpointTest.php
@@ -12,6 +12,10 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint('https://laravel.com', 'docs');
 
+        $this->assertSame('https://laravel.com', $endpoint->getUri());
+        $this->assertSame(['docs'], $endpoint->getPath());
+        $this->assertSame([], $endpoint->getQuery());
+
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs', (string) $endpoint->get());
     }
@@ -20,6 +24,10 @@ class EndpointTest extends TestCase
     function it_can_build_basic_endpoint_with_tailing_slash()
     {
         $endpoint = new Endpoint('https://laravel.com/', 'docs');
+
+        $this->assertSame('https://laravel.com', $endpoint->getUri());
+        $this->assertSame(['docs'], $endpoint->getPath());
+        $this->assertSame([], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs', (string) $endpoint->get());
@@ -30,6 +38,10 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint('https://laravel.com', ['docs', '5.4']);
 
+        $this->assertSame('https://laravel.com', $endpoint->getUri());
+        $this->assertSame(['docs', '5.4'], $endpoint->getPath());
+        $this->assertSame([], $endpoint->getQuery());
+
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs/5.4', (string) $endpoint->get());
     }
@@ -38,6 +50,10 @@ class EndpointTest extends TestCase
     function it_can_build_basic_endpoint_with_query_string()
     {
         $endpoint = new Endpoint('https://laravel.com', 'docs', ['search' => 'controller']);
+
+        $this->assertSame('https://laravel.com', $endpoint->getUri());
+        $this->assertSame(['docs'], $endpoint->getPath());
+        $this->assertSame(['search' => 'controller'], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs?search=controller', (string) $endpoint->get());
@@ -49,7 +65,24 @@ class EndpointTest extends TestCase
         $endpoint = (new Endpoint('https://laravel.com', 'docs'))
                         ->addQuery(['search' => 'controller', 'page' => 3]);
 
+        $this->assertSame('https://laravel.com', $endpoint->getUri());
+        $this->assertSame(['docs'], $endpoint->getPath());
+        $this->assertSame(['search' => 'controller', 'page' => 3], $endpoint->getQuery());
+
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs?search=controller&page=3', (string) $endpoint->get());
+    }
+
+    /** @test */
+    function it_can_set_endpoint_with_no_uri()
+    {
+        $endpoint = new Endpoint(null, 'docs');
+
+        $this->assertSame(null, $endpoint->getUri());
+        $this->assertSame(['docs'], $endpoint->getPath());
+        $this->assertSame([], $endpoint->getQuery());
+
+        $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
+        $this->assertSame('/docs', (string) $endpoint->get());
     }
 }