+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 This changelog references the relevant changes (bug and security fixes) done to `laravie/codex`.
 
+## 1.0.1
+
+Released: Not yet
+
+### Changes
+
+* Allow `Laravie\Codex\Request::send()` to properly build URI if `$path` is given as `Laravie\Codex\Endpoint`.
 
 ## 1.0.0
 
-Released: `<YYYY-MM-DD>`
+Released: 2017-08-31
 
 ### New
 
@@ -29,6 +36,7 @@ Released: `<YYYY-MM-DD>`
 * Move declaration `sanitizeWith()` from `Laravie\Codex\Client` to `Request` class, ignore this if sanitization isn't needed.
 * `Laravie\Codex\Request` should set sanitizer to `Laravie\Codex\Response` before running `Response::validate()` method.
 * Improves sanitization supports.
+* Body shouldn't be converted to query string if it's an instance of `StreamInterface`.
 
 ### Removed